Resolved issue where lookup form would display for already populated address fields (e.g. an existing node).
Updated the lookup function of the pca predirect service to append the next operation to the ID. Colon separated.

Added some enum values to the PCA Predict abstract class for indicating 'Next' operations.

Upated the retrieve function of the API adapter to split the ID value on a colon to get the actual ID and 'Next' operation.

Refactored multi-step lookup logic to support result sets with a mix of 'Find' and 'Next' operations.

Added custom AJAX callback for the address select drop down for use in PCA Predict multi-step lookups.